add isLoading toggle to localStorage
