+++ client/src/context/userContext.js
@@ -5,11 +5,21 @@ export const UserContext = React.createContext();
 
 // create context provider
 export const UserProvider = ({ children }) => {
+  const [isLoading, setIsLoading] = useState(false);
   const [userToken, setUserToken] = useState('default user context');
   const [user, setUser] = useState('default user object');
   return (
     // inject state into the provider, and pass along to children components
-    <UserContext.Provider value={{ user, setUser, userToken, setUserToken }}>
+    <UserContext.Provider
+      value={{
+        user,
+        setUser,
+        userToken,
+        setUserToken,
+        isLoading,
+        setIsLoading,
+      }}
+    >
       {children}
     </UserContext.Provider>
   );
